+++ packages/manager/src/features/Longview/LongviewDetail/DetailTabs/Apache/Apache.tsx
@@ -17,7 +17,7 @@ const useStyles = makeStyles((theme: Theme) => ({
   root: {
     width: 250,
   },
-  cmrSpacing: {
+  title: {
     [theme.breakpoints.down('md')]: {
       marginLeft: theme.spacing(),
       marginRight: theme.spacing(),
@@ -102,8 +102,8 @@ export const Apache: React.FC<Props> = (props) => {
           alignItems="center"
         >
           <div>
-            <Typography className={classes.cmrSpacing} variant="h2">
-              {'Apache'}
+            <Typography className={classes.title} variant="h2">
+              Apache
             </Typography>
             {version && <Typography variant="body1">{version}</Typography>}
           </div>
