+++ packages/manager/src/features/Longview/LongviewDetail/DetailTabs/MySQL/MySQLLanding.tsx
@@ -17,7 +17,7 @@ const useStyles = makeStyles((theme: Theme) => ({
   root: {
     width: 250,
   },
-  cmrSpacing: {
+  title: {
     [theme.breakpoints.down('md')]: {
       marginLeft: theme.spacing(),
       marginRight: theme.spacing(),
@@ -103,7 +103,7 @@ export const MySQLLanding: React.FC<Props> = (props) => {
           alignItems="center"
         >
           <div>
-            <Typography className={classes.cmrSpacing} variant="h2">
+            <Typography className={classes.title} variant="h2">
               MySQL
             </Typography>
             {version && <Typography variant="body1">{version}</Typography>}
