+++ packages/manager/src/features/Support/SupportTicketDetail/TabbedReply/TabbedReply.tsx
@@ -1,16 +1,9 @@
 import * as React from 'react';
 import { compose } from 'recompose';
-import {
-  createStyles,
-  Theme,
-  withStyles,
-  WithStyles,
-} from 'src/components/core/styles';
+import { makeStyles } from 'src/components/core/styles';
 import TabbedPanel, { Tab } from 'src/components/TabbedPanel';
 import Preview from './PreviewReply';
 import Reply, { Props as ReplyProps } from './TicketReply';
-
-type ClassNames = 'root';
 
 interface Props {
   rootClass?: string;
@@ -19,17 +12,20 @@ interface Props {
   required?: boolean;
 }
 
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {
-      backgroundColor: 'transparent',
-    },
-  });
+const useStyles = makeStyles(() => ({
+  root: {
+    backgroundColor: 'transparent',
+    paddingRight: 0,
+    paddingLeft: 0,
+  },
+}));
 
-type CombinedProps = Props & ReplyProps & WithStyles<ClassNames>;
+type CombinedProps = Props & ReplyProps;
 
 const TabbedReply: React.FC<CombinedProps> = (props) => {
-  const { innerClass, rootClass, classes, value, error, ...rest } = props;
+  const classes = useStyles();
+
+  const { innerClass, rootClass, value, error, ...rest } = props;
 
   const title = props.isReply ? 'Reply' : 'Description';
 
@@ -59,8 +55,6 @@ const TabbedReply: React.FC<CombinedProps> = (props) => {
   );
 };
 
-const styled = withStyles(styles);
-
 /** only update on error and value change */
 const memoized = (component: React.FC<CombinedProps>) =>
   React.memo<CombinedProps>(component, (prevProps, nextProps) => {
@@ -71,7 +65,6 @@ const memoized = (component: React.FC<CombinedProps>) =>
     );
   });
 
-export default compose<CombinedProps, ReplyProps & Props>(
-  memoized,
-  styled
-)(TabbedReply);
+export default compose<CombinedProps, ReplyProps & Props>(memoized)(
+  TabbedReply
+);
