+++ packages/manager/src/features/linodes/LinodesDetail/LinodeResize/LinodeResize.tsx
@@ -8,15 +8,13 @@ import {
 import { APIError } from '@linode/api-v4/lib/types';
 import { withSnackbar, WithSnackbarProps } from 'notistack';
 import * as React from 'react';
-import { connect, MapDispatchToProps } from 'react-redux';
-import { RouteComponentProps, withRouter } from 'react-router-dom';
+import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux';
 import { compose } from 'recompose';
 import { Action } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
 import ActionsPanel from 'src/components/ActionsPanel';
 import Button from 'src/components/Button';
 import Checkbox from 'src/components/CheckBox';
-import Paper from 'src/components/core/Paper';
 import {
   createStyles,
   Theme,
@@ -24,21 +22,27 @@ import {
   WithStyles,
 } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
-import { DocumentTitleSegment } from 'src/components/DocumentTitle';
+import Dialog from 'src/components/Dialog';
 import ExternalLink from 'src/components/ExternalLink';
 import HelpIcon from 'src/components/HelpIcon';
+import Notice from 'src/components/Notice';
+import TextField from 'src/components/TextField';
 import withTypes, { WithTypesProps } from 'src/containers/types.container';
 import { resetEventsPolling } from 'src/eventsPolling';
 import SelectPlanPanel from 'src/features/linodes/LinodesCreate/SelectPlanPanel';
-import { ExtendedType } from 'src/store/linodeType/linodeType.reducer';
-import { withLinodeDetailContext } from 'src/features/linodes/LinodesDetail/linodeDetailContext';
 import { linodeInTransition } from 'src/features/linodes/transitions';
 import { ApplicationState } from 'src/store';
+import { getAllLinodeDisks } from 'src/store/linodes/disk/disk.requests';
+import { getLinodeDisksForLinode } from 'src/store/linodes/disk/disk.selectors';
 import { requestLinodeForStore } from 'src/store/linodes/linode.requests';
+import { getPermissionsForLinode } from 'src/store/linodes/permissions/permissions.selector';
+import { ExtendedType } from 'src/store/linodeType/linodeType.reducer';
+import { EntityError } from 'src/store/types';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
+import { GetAllData } from 'src/utilities/getAll';
+import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
 import HostMaintenanceError from '../HostMaintenanceError';
 import LinodePermissionsError from '../LinodePermissionsError';
-import ResizeConfirmation from './ResizeConfirmationDialog';
 
 type ClassNames =
   | 'root'
@@ -47,28 +51,11 @@ type ClassNames =
   | 'toolTip'
   | 'currentPlanContainer'
   | 'resizeTitle'
-  | 'checkbox'
   | 'currentHeaderEmptyCell'
-  | 'actions'
-  | 'errorLink';
+  | 'selectPlanPanel';
 
 const styles = (theme: Theme) =>
   createStyles({
-    root: {
-      padding: theme.spacing(3),
-      paddingBottom: theme.spacing(2),
-      '& .tabbedPanel': {
-        '& > div': {
-          padding: 0,
-        },
-      },
-    },
-    checkbox: {
-      marginTop: theme.spacing(3),
-      '& .MuiButtonBase-root': {
-        marginLeft: 3,
-      },
-    },
     toolTip: {
       paddingTop: theme.spacing(1),
     },
@@ -92,93 +79,51 @@ const styles = (theme: Theme) =>
     currentHeaderEmptyCell: {
       width: '13%',
     },
-    actions: {
-      paddingBottom: theme.spacing(2),
-      paddingLeft: theme.spacing(3),
-    },
-    errorLink: {
-      color: '#c44742',
-      textDecoration: 'underline',
+    selectPlanPanel: {
+      marginTop: theme.spacing(5),
+      marginBottom: theme.spacing(3),
+      '& > div': {
+        padding: 0,
+      },
     },
   });
 
-interface LinodeContextProps {
-  linodeId: number;
-  linodeType: null | string;
-  linodeStatus?: LinodeStatus;
-  linodeLabel: string;
-  permissions: GrantLevel;
-  linodeDisks: Disk[];
-}
-
-interface ConfirmationDialog {
-  isOpen: boolean;
-  error?: string | JSX.Element;
-  submitting: boolean;
-  currentPlan: string;
-  targetPlan: string;
+interface Props {
+  linodeId?: number;
+  linodeLabel?: string;
+  open: boolean;
+  onClose: () => void;
 }
 
 interface State {
   selectedId: string;
   errors?: APIError[];
   autoDiskResize: boolean;
-  confirmationDialog: ConfirmationDialog;
+  confirmationText: string;
+  submitting: boolean;
+  submissionError?: string | JSX.Element;
 }
 
-type CombinedProps = WithTypesProps &
-  RouteComponentProps &
-  LinodeContextProps &
+type CombinedProps = Props &
+  WithTypesProps &
   WithStyles<ClassNames> &
   DispatchProps &
-  WithSnackbarProps;
+  WithSnackbarProps &
+  StateProps;
 
 export class LinodeResize extends React.Component<CombinedProps, State> {
   state: State = {
     selectedId: '',
-    autoDiskResize: shouldEnableAutoResizeDiskOption(this.props.linodeDisks)[1],
-    confirmationDialog: {
-      isOpen: false,
-      submitting: false,
-      currentPlan: '',
-      targetPlan: '',
-    },
-  };
-
-  openConfirmationModal = () => {
-    const { linodeType, typesData } = this.props;
-    const { selectedId } = this.state;
-    const _current = getLinodeType(typesData, linodeType || 'none');
-    const _target = getLinodeType(typesData, selectedId);
-    const currentPlan = _current ? _current.label : 'Unknown plan';
-    const targetPlan = _target ? _target.label : 'Unknown plan';
-    this.setState({
-      confirmationDialog: {
-        ...this.state.confirmationDialog,
-        isOpen: true,
-        error: undefined,
-        currentPlan,
-        targetPlan,
-      },
-    });
-  };
-
-  closeConfirmationModal = () => {
-    this.setState({
-      confirmationDialog: {
-        ...this.state.confirmationDialog,
-        isOpen: false,
-      },
-    });
+    autoDiskResize: false,
+    confirmationText: '',
+    submitting: false,
   };
 
   onSubmit = () => {
     const {
-      classes,
       linodeId,
       linodeType,
       enqueueSnackbar,
-      history,
       updateLinode,
       typesData,
     } = this.props;
@@ -195,10 +140,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
     );
 
     this.setState({
-      confirmationDialog: {
-        ...this.state.confirmationDialog,
-        submitting: true,
-      },
+      submitting: true,
     });
 
     /**
@@ -211,11 +153,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
       .then((_) => {
         this.setState({
           selectedId: '',
-          confirmationDialog: {
-            ...this.state.confirmationDialog,
-            submitting: false,
-            isOpen: false,
-          },
+          submitting: false,
         });
         resetEventsPolling();
         enqueueSnackbar('Linode queued for resize.', {
@@ -228,28 +166,38 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
         // go away.
         updateLinode(linodeId);
 
-        history.push(`/linodes/${linodeId}/summary`);
+        this.props.onClose();
       })
       .catch((errorResponse) => {
         let error: string | JSX.Element = '';
         const reason = errorResponse[0]?.reason ?? '';
+        /**
+         * The logic below is to manually intercept a certain
+         * error and add some JSX with a hyperlink to it.
+         *
+         * Unfortunately, we have global error interceptors that
+         * do the same thing, as is the case when your reputation
+         * score is too low to do what you're trying to do.
+         *
+         * If one of those already-intercepted errors comes through here,
+         * it will break the logic (since error[0].reason is not a string).
+         */
         if (
           typeof reason === 'string' &&
           reason.match(/allocated more disk/i)
         ) {
           error = (
-            <>
+            <Typography>
               The current disk size of your Linode is too large for the new
               service plan. Please resize your disk to accommodate the new plan.
               You can read our{' '}
               <ExternalLink
-                className={classes.errorLink}
                 hideIcon
                 text="Resize Your Linode"
                 link="https://www.linode.com/docs/platform/disk-images/resizing-a-linode/"
               />{' '}
               guide for more detailed instructions.
-            </>
+            </Typography>
           );
         } else {
           error = getAPIErrorOrDefault(
@@ -257,14 +205,12 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
             'There was an issue resizing your Linode.'
           )[0].reason;
         }
-
         this.setState({
-          confirmationDialog: {
-            ...this.state.confirmationDialog,
-            error,
-            submitting: false,
-          },
+          submissionError: error,
+          submitting: false,
         });
+        // Set to "block: end" since the sticky header would otherwise interfere.
+        scrollErrorIntoView(undefined, { block: 'end' });
       });
   };
 
@@ -276,22 +222,51 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
     this.setState({ autoDiskResize: !this.state.autoDiskResize });
   };
 
+  componentDidUpdate = (prevProps: CombinedProps) => {
+    if (
+      prevProps.linodeId !== this.props.linodeId &&
+      !!this.props.linodeId &&
+      this.props.linodeDisks?.length === 0
+    ) {
+      // @todo: Error?
+      this.props.getLinodeDisks(this.props.linodeId).catch((_) => null);
+    }
+
+    if (
+      prevProps.linodeDisks?.length == 0 &&
+      this.props.linodeDisks && // TS Compiler wasn't recognizing the optional chaining here.
+      this.props.linodeDisks.length > 0
+    ) {
+      this.setState({
+        autoDiskResize: shouldEnableAutoResizeDiskOption(
+          this.props.linodeDisks ?? []
+        )[1],
+      });
+    }
+  };
+
   render() {
     const {
       typesData,
       linodeType,
-      linodeLabel,
       permissions,
       classes,
       linodeDisks,
       linodeStatus,
+      linodeDisksError,
+      linodeLabel,
     } = this.props;
+    const { confirmationText, submissionError } = this.state;
     const type = typesData.find((t) => t.id === linodeType);
 
     const hostMaintenance = linodeStatus === 'stopped';
     const unauthorized = permissions === 'read_only';
+    const disksError = linodeDisksError?.read;
 
-    const disabled = hostMaintenance || unauthorized;
+    const tableDisabled =
+      hostMaintenance || unauthorized || Boolean(disksError);
+
+    const submitButtonDisabled = confirmationText !== linodeLabel;
 
     const currentPlanHeading = linodeType
       ? type
@@ -302,7 +277,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
     const [
       diskToResize,
       _shouldEnableAutoResizeDiskOption,
-    ] = shouldEnableAutoResizeDiskOption(linodeDisks);
+    ] = shouldEnableAutoResizeDiskOption(linodeDisks ?? []);
 
     const isSmaller = isSmallerThanCurrentPlan(
       this.state.selectedId,
@@ -311,32 +286,35 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
     );
 
     return (
-      <div>
-        <DocumentTitleSegment segment={`${linodeLabel} - Resize`} />
-        <Paper className={classes.root}>
-          {unauthorized && <LinodePermissionsError />}
-          {hostMaintenance && <HostMaintenanceError />}
-          <Typography
-            role="heading"
-            aria-level={2}
-            variant="h2"
-            className={classes.title}
-            data-qa-title
-          >
-            Resize
-          </Typography>
-          <Typography data-qa-description>
-            If you&apos;re expecting a temporary burst of traffic to your
-            website, or if you&apos;re not using your Linode as much as you
-            thought, you can temporarily or permanently resize your Linode to a
-            different plan.{' '}
-            <ExternalLink
-              fixedIcon
-              text="Learn more."
-              link="https://www.linode.com/docs/platform/disk-images/resizing-a-linode/"
-            />
-          </Typography>
+      <Dialog
+        title={`Resize Linode ${this.props.linodeLabel ?? ''}`}
+        open={this.props.open}
+        onClose={this.props.onClose}
+        fullWidth
+        fullHeight
+        maxWidth="md"
+      >
+        {unauthorized && <LinodePermissionsError />}
+        {hostMaintenance && <HostMaintenanceError />}
+        {disksError && (
+          <Notice
+            error
+            text="There was an error loading your Linode's Disks."
+          />
+        )}
+        {submissionError && <Notice error>{submissionError}</Notice>}
+        <Typography data-qa-description>
+          If you&apos;re expecting a temporary burst of traffic to your website,
+          or if you&apos;re not using your Linode as much as you thought, you
+          can temporarily or permanently resize your Linode to a different plan.{' '}
+          <ExternalLink
+            fixedIcon
+            text="Learn more."
+            link="https://www.linode.com/docs/platform/disk-images/resizing-a-linode/"
+          />
+        </Typography>
 
+        <div className={classes.selectPlanPanel}>
           <SelectPlanPanel
             currentPlanHeading={currentPlanHeading}
             types={typesData.filter(
@@ -344,70 +322,85 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
             )}
             onSelect={this.handleSelectPlan}
             selectedID={this.state.selectedId}
-            disabled={disabled}
+            disabled={tableDisabled}
+            updateFor={[this.state.selectedId]}
           />
-        </Paper>
-        <Paper className={`${classes.checkbox} ${classes.root}`}>
-          <Typography variant="h2" className={classes.resizeTitle}>
-            Auto Resize Disk
-            {isSmaller ? (
-              <HelpIcon
-                className={classes.toolTip}
-                text={`Your disks cannot be automatically resized when moving to a smaller plan.`}
-              />
-            ) : !_shouldEnableAutoResizeDiskOption ? (
-              <HelpIcon
-                className={classes.toolTip}
-                text={`Your ext disk can only be automatically resized if you have one ext
+        </div>
+        <Typography variant="h2" className={classes.resizeTitle}>
+          Auto Resize Disk
+          {disksError ? (
+            <HelpIcon
+              className={classes.toolTip}
+              text={`There was an error loading your Linode's disks.`}
+            />
+          ) : isSmaller ? (
+            <HelpIcon
+              className={classes.toolTip}
+              text={`Your disks cannot be automatically resized when moving to a smaller plan.`}
+            />
+          ) : !_shouldEnableAutoResizeDiskOption ? (
+            <HelpIcon
+              className={classes.toolTip}
+              text={`Your ext disk can only be automatically resized if you have one ext
                       disk or one ext disk and one swap disk on this Linode.`}
-              />
-            ) : null}
+            />
+          ) : null}
+        </Typography>
+        <Checkbox
+          disabled={!_shouldEnableAutoResizeDiskOption || isSmaller}
+          checked={
+            !_shouldEnableAutoResizeDiskOption || isSmaller
+              ? false
+              : this.state.autoDiskResize
+          }
+          onChange={this.handleToggleAutoDisksResize}
+          text={
+            <Typography>
+              Would you like{' '}
+              {_shouldEnableAutoResizeDiskOption ? (
+                <strong>{diskToResize}</strong>
+              ) : (
+                'your disk'
+              )}{' '}
+              to be automatically scaled with this Linode&apos;s new size? We
+              recommend you keep this option enabled when available. Automatic
+              resizing is only available when moving to a larger plan, and when
+              you have a single ext disk (or one ext and one swap disk) on your
+              Linode.
+            </Typography>
+          }
+        />
+
+        <ActionsPanel>
+          <Typography variant="h2">Confirm</Typography>
+          <Typography style={{ marginBottom: 8 }}>
+            To confirm these changes, type the label of the Linode{' '}
+            <strong>({linodeLabel})</strong> in the field below:
           </Typography>
-          <Checkbox
-            disabled={!_shouldEnableAutoResizeDiskOption || isSmaller}
-            checked={
-              !_shouldEnableAutoResizeDiskOption || isSmaller
-                ? false
-                : this.state.autoDiskResize
+          <TextField
+            label="Linode Label"
+            hideLabel
+            onChange={(e) =>
+              this.setState({ confirmationText: e.target.value })
             }
-            onChange={this.handleToggleAutoDisksResize}
-            text={
-              <Typography>
-                Would you like{' '}
-                {_shouldEnableAutoResizeDiskOption ? (
-                  <strong>{diskToResize}</strong>
-                ) : (
-                  'your disk'
-                )}{' '}
-                to be automatically scaled with this Linode&apos;s new size? We
-                recommend you keep this option enabled when available. Automatic
-                resizing is only available when moving to a larger plan, and
-                when you have a single ext disk (or one ext and one swap disk)
-                on your Linode.
-              </Typography>
-            }
+            style={{ marginBottom: 16 }}
           />
-        </Paper>
-        <ActionsPanel className={classes.actions}>
           <Button
             disabled={
               !this.state.selectedId ||
               linodeInTransition(this.props.linodeStatus || '') ||
-              disabled
+              tableDisabled ||
+              submitButtonDisabled
             }
+            loading={this.state.submitting}
             buttonType="primary"
-            onClick={this.openConfirmationModal}
+            onClick={this.onSubmit}
             data-qa-resize
           >
-            Resize
+            Resize Linode
           </Button>
         </ActionsPanel>
-        <ResizeConfirmation
-          {...this.state.confirmationDialog}
-          onClose={this.closeConfirmationModal}
-          onResize={this.onSubmit}
-        />
-      </div>
+      </Dialog>
     );
   }
 }
@@ -416,6 +409,7 @@ const styled = withStyles(styles);
 
 interface DispatchProps {
   updateLinode: (id: number) => void;
+  getLinodeDisks: (linodeId: number) => Promise<GetAllData<Disk>>;
 }
 
 const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = (
@@ -423,22 +417,51 @@ const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = (
 ) => {
   return {
     updateLinode: (id: number) => dispatch(requestLinodeForStore(id)),
+    getLinodeDisks: (linodeId: number) =>
+      dispatch(getAllLinodeDisks({ linodeId })),
   };
 };
 
-const connected = connect(undefined, mapDispatchToProps);
+interface StateProps {
+  linodeId?: number;
+  linodeType?: null | string;
+  linodeStatus?: LinodeStatus;
+  linodeLabel?: string;
+  permissions?: GrantLevel;
+  linodeDisks?: Disk[];
+  linodeDisksError?: EntityError;
+}
 
-const linodeContext = withLinodeDetailContext((state) => {
-  const { linode } = state;
+const mapStateToProps: MapStateToProps<StateProps, Props> = (
+  state: ApplicationState,
+  ownProps
+) => {
+  const { linodeId } = ownProps;
+
+  if (!linodeId) {
+    return {};
+  }
+
+  const linode = state.__resources.linodes.itemsById[linodeId];
+  const linodeDisks = state.__resources.linodeDisks;
+  const profile = state.__resources.profile;
+
+  if (!linode) {
+    return {};
+  }
+
   return {
     linodeId: linode.id,
     linodeType: linode.type,
     linodeStatus: linode.status,
     linodeLabel: linode.label,
-    permissions: linode._permissions,
-    linodeDisks: linode._disks,
+    permissions: getPermissionsForLinode(profile.data ?? null, linodeId),
+    linodeDisks: getLinodeDisksForLinode(linodeDisks, linodeId),
+    linodeDisksError: linodeDisks[linodeId]?.error,
   };
-});
+};
+
+const connected = connect(mapStateToProps, mapDispatchToProps);
 
 export const getLinodeType = (types: LinodeType[], selectedTypeID: string) => {
   return types.find((thisType) => thisType.id === selectedTypeID);
@@ -494,11 +517,9 @@ export const isSmallerThanCurrentPlan = (
   return currentType.disk > nextType.disk;
 };
 
-export default compose<CombinedProps, {}>(
-  linodeContext,
+export default compose<CombinedProps, Props>(
   withTypes,
   styled,
   withSnackbar,
-  withRouter,
   connected
 )(LinodeResize);
